fix overflow in size calculation

+++ include/dedispersionliteU.h
@@ -83,10 +83,10 @@ namespace Pulsar
 	public:
 		long int counter;
 		long int offset;
-		int nchans;
+		long int nchans;
 		double tsamp;
-		int ndump;
-		int nsamples;
+		long int ndump;
+		long int nsamples;
 		vector<double> frequencies;
 		vector<long int> delayn;
 		vector<float> buffer;
